Move PitchedMalloc for commonality
